requirements.txt
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
gino==1.0.1
hiredis==1.0.1
idna==2.10
importlib-metadata==1.7.0
multidict==4.7.6
python-dotenv==0.14.0
pytz==2020.1
SQLAlchemy==1.3.18
yarl==1.4.2
zipp==3.1.0
